+++ src/main/java/muramasa/antimatter/tile/multi/TileEntityMultiMachine.java
@@ -242,7 +242,7 @@ public class TileEntityMultiMachine extends TileEntityMachine implements ICompon
     }
 
     @Override
-    public long getMaxInputVoltage() {
+    public int getMaxInputVoltage() {
         List<IComponentHandler> hatches = getComponents("hatch_energy");
         return hatches.size() >= 1 ? hatches.get(0).getEnergyHandler().isPresent() ? hatches.get(0).getEnergyHandler().get().getInputVoltage() : Ref.V[0] : Ref.V[0];
     }
